refactor: split screenshot and hot-reload patcher sources below the file-length limit
Keep ScreenshotUseCase SwitchTo sites in-method and HotReloadPatcher ledgers
on the patcher so the 500 SLOC gate can land without changing capture or
transplant behavior.
