+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadPatchIlRebind.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 414f2b78dd80d4532ac394231eaad503
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
