+++ Packages/src/Editor/FirstPartyTools/Screenshot/ScreenshotCaptureResults.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6845236c3a3014d7593c471fdf088a8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
