+++ Packages/src/Editor/FirstPartyTools/Screenshot/ScreenshotOverlayControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7f86124e068047799428eb814a9558a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
