+++ Packages/src/Editor/FirstPartyTools/Screenshot/ScreenshotRaycastGridCollector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af44fc5c7db9b478983cfcbcc0c7a5ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
